interpol and interpolate, separated, clean
