+++ src/interpolate.hpp
@@ -1,0 +1,34 @@
+/***************************************************************************
+     interpolate.hpp  -  all things related to interpolate command
+                             -------------------
+    begin                : Mar 30 2021
+    copyright            : (C) 2004 by Joel Gales
+                         : (C) 2018 G. Duvert 
+    email                : see https://github.com/gnudatalanguage/gdl
+ ***************************************************************************/
+
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+
+#ifdef HAVE_CONFIG_H
+#include <config.h>
+#endif
+
+#if defined(HAVE_LIBGSL)
+
+#include "datatypes.hpp"
+#include "envt.hpp"
+
+namespace lib {
+
+  BaseGDL* interpolate_fun( EnvT* e);
+
+} // namespace
+
+#endif
