+++ src/templates/new.cpp
@@ -1,0 +1,24 @@
+/***************************************************************************
+                          |FILENAME|  -  description
+                             -------------------
+    begin                : |DATE|
+    copyright            : (C) |YEAR| by |AUTHOR|
+    email                : |EMAIL|
+ ***************************************************************************/
+
+/***************************************************************************
+ *                                                                         *
+ *   This program is free software; you can redistribute it and/or modify  *
+ *   it under the terms of the GNU General Public License as published by  *
+ *   the Free Software Foundation; either version 2 of the License, or     *
+ *   (at your option) any later version.                                   *
+ *                                                                         *
+ ***************************************************************************/
+
+#include "includefirst.hpp"
+#include "datatypes.hpp"
+#include "envt.hpp"
+#include "dinterpreter.hpp"
+
+namespace lib {
+}
