+++ src/templates/new.hpp
